requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Experiment tracking
mlflow>=2.9.0

# API
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0

# Data versioning
dvc>=3.30.0

# Pipeline
kfp>=2.4.0  # Kubeflow Pipelines SDK

# Monitoring
evidently>=0.4.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# GenAI / synthetic data
langchain>=0.1.0
openai>=1.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # for FastAPI test client

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
